Move plugin/ → plugins/omh/ to match upstream convention
Replaces upstream's basic implementation with our security-hardened
version at the same path. Updates all imports, test patches, runner
scripts, and skill references.

93 unit tests + 9 Hermes venv integration tests pass.
