--- plugin/config.yaml
@@ -1,72 +1,0 @@
-role_prompts_dir: ~/.hermes/skills/omh-ralplan/references
-
-roles:
-  executor:          {category: implementation}
-  verifier:          {category: review}
-  architect:         {category: analysis}
-  planner:           {category: planning}
-  critic:            {category: analysis}
-  analyst:           {category: analysis}
-  security-reviewer: {category: review}
-  code-reviewer:     {category: review}
-  test-engineer:     {category: testing}
-  debugger:          {category: analysis}
-
-state_dir: .omh/state
-staleness_hours: 2
-cancel_ttl_seconds: 30
-
-evidence:
-  # Allowlist controls which commands omh_gather_evidence may run.
-  # Each entry is a token-prefix match: the leading tokens of the command must exactly
-  # equal the entry's tokens. E.g. "npm test" matches "npm test --verbose" but NOT
-  # "npm testing-malicious". Add entries here to expand what the agent can verify.
-  allowlist_prefixes:
-    # npm — specific test/build/lint subcommands only (not "npm " which permits npm publish, npm install, etc.)
-    - "npm test"
-    - "npm run test"
-    - "npm run build"
-    - "npm run lint"
-    - "npm run typecheck"
-    - "npm run check"
-    # npx — specific trusted runners only (not "npx " which permits arbitrary package execution)
-    - "npx jest"
-    - "npx vitest"
-    - "npx mocha"
-    - "npx eslint"
-    - "npx tsc"
-    # yarn / pnpm
-    - "yarn test"
-    - "yarn build"
-    - "yarn lint"
-    - "pnpm test"
-    - "pnpm build"
-    - "pnpm lint"
-    # Rust
-    - "cargo test"
-    - "cargo build"
-    - "cargo check"
-    - "cargo clippy"
-    # Go
-    - "go test"
-    - "go build"
-    - "go vet"
-    # Python
-    - "python -m pytest"
-    - "python -m mypy"
-    - "python3 -m pytest"
-    - "python3 -m mypy"
-    # Linters / formatters
-    - "ruff check"
-    - "ruff format --check"
-    - "black --check"
-    - "eslint"
-    - "tsc"
-    # Make — specific safe targets only
-    - "make test"
-    - "make build"
-    - "make check"
-    - "make lint"
-  max_commands: 10
-  default_timeout: 120
-  default_truncate: 2000
