--- plugin/plugin.yaml
@@ -1,16 +1,0 @@
-name: omh
-version: "0.1.0"
-description: "Oh My Hermes — infrastructure layer for OMH skills (state management, evidence gathering, mode awareness)"
-entry_point: plugin
-requires_python: ">=3.10"
-
-tools:
-  - omh_state
-  - omh_gather_evidence
-
-hooks:
-  - pre_llm_call
-  - on_session_end
-
-optional_dependencies:
-  - pyyaml  # required for config.yaml loading; falls back to built-in defaults if absent
