+++ plugins/omh/hooks/llm_hooks.py
