+++ plugins/omh/tests/run_integration.sh
@@ -2,12 +2,12 @@
 # Run integration tests using the Hermes venv Python.
 # This validates the plugin against the actual Hermes runtime.
 #
-# Usage: ./plugin/tests/run_integration.sh
+# Usage: ./plugins/omh/tests/run_integration.sh
 #
 set -euo pipefail
 
 HERMES_VENV="${HOME}/.hermes/hermes-agent/venv/bin/python3"
-REPO_ROOT="$(cd "$(dirname "$0")/../.." && pwd)"
+REPO_ROOT="$(cd "$(dirname "$0")/../../.." && pwd)"
 
 if [ ! -x "$HERMES_VENV" ]; then
   echo "ERROR: Hermes venv not found at $HERMES_VENV"
@@ -17,4 +17,4 @@ fi
 
 echo "Running integration tests with Hermes venv: $HERMES_VENV"
 cd "$REPO_ROOT"
-PYTHONPATH="$REPO_ROOT" exec "$HERMES_VENV" plugin/tests/integration_runner.py "$@"
+PYTHONPATH="$REPO_ROOT" exec "$HERMES_VENV" plugins/omh/tests/integration_runner.py "$@"
